Use lodash-es for improved ES Module support
Swap lodash for lodash-es and import only the one module we need,
instead of importing the entire library and then referencing the module
as a method on the library.

lodash-es provides better resolution into the individual lodash modules
which make it into the final bundle for both webpack-bundle-analyzer and
source-map-explorer.

Explicitly importing only the range method allows webpack to bundle only
the lodash modules which we need in order to use range. Some webpack and
babel plugins endeavor to accomplish this automatically, which is
definitely worth exploring:
https://www.npmjs.com/package/babel-plugin-lodash and
https://www.npmjs.com/package/lodash-webpack-plugin